Move WalletDAOs to testkit
